refactor: AuthenticationException 상속 처리
