--- src/main/java/subway/auth/AuthenticationException.java
@@ -1,8 +1,0 @@
-package subway.auth;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class AuthenticationException extends RuntimeException {
-}
