+++ src/main/java/subway/auth/token/TokenService.java
@@ -1,6 +1,6 @@
 package subway.auth.token;
 
-import subway.auth.AuthenticationException;
+import subway.exception.AuthenticationException;
 import subway.auth.token.oauth2.OAuth2User;
 import subway.auth.token.oauth2.OAuth2UserService;
 import subway.auth.token.oauth2.github.GithubClient;
@@ -31,7 +31,7 @@ public class TokenService {
     public TokenResponse createToken(String email, String password) {
         UserDetails userDetails = userDetailsService.loadUserByUsername(email);
         if (!userDetails.getPassword().equals(password)) {
-            throw new AuthenticationException();
+            throw new AuthenticationException(9999L, "로그인 정보가 일치하지 않습니다."); // TODO: constant
         }
 
         String token = jwtTokenProvider.createToken(userDetails.getUsername(), userDetails.getRole());
